Add MailleInlay struct
Add struct that contains all variables related to the design of the
inlay itself. Contains all the data necessary to display the inlay
design so it can be passed around to various tools, giving them full
control to modify the design without giving them access to unrelated
settings and variables.