Decrease shared_ptr usage and add default tool functions
The inlay object does not need to be a shared pointer so we should pass
it to the tools by reference rather than throwing shared_ptr objects all
around. Also there is no reason not to have a default do-nothing
implementation for tool virtual functions, so implement those as well.