Add typedef for Color type
Defines a color to be a vector of 3 unsigned chars. This will allow
colors to be considered strictly equal/unequal to each other so we can
later do reporting on how many rings of each color are necessary. May
also make serializing the data a little nicer since it is integers
rather than floating point numbers.