Improve rounding in color wheel callback
Using rounding instead of the floor operation previously used allows
colors to be more consistently recognized as the same color used
elsewhere in the project. Without this fix, if a color is selected using
the color picker tool and then used for a new ring, it is not uncommon
to see the new ring be recognized as a different color in the ring color
histogram due to rounding issues.