Remove ability to rotate the view matrix
Now that we have explicit support for wrongWay/rightWay in the
European-4-in-1 weave, we don't need to rotate the whole view matrix to
do inlays the Right Way.

This has the added benefit that rotating and then resizing the window is
no longer possible, and thus not a bug.