Add custom button to allow tools to have popups
This adds a modified PopupButton which allows the user to easily disable
the popup altogether. Many tools do not need (and therefore do not want)
the popups from the PopupButton, so having the ability to disable the
popup is very useful. This commit moves the selection delete and
selection set color tools to a popup connected to the selection tool
button.