Fix bug in index calculation for imported images
Due to the 4 in 1 grid index pattern used, there are some indexes which
are not supposed to exist in the pattern. For example, index (0, 1) is
not normally able to exist, but due to the image import implementation,
this index would sometimes be created when importing images. This commit
fixes the y indices such that the indices created by importing an image
stay on the expected grid, which prevents rings getting added by the
other ring adding tools where they don't actually fit into the weave.